+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This release will bump the Realm file format from version 22 to 23. Opening a fi
 * Realm will now continuously track and reduce the size of the Realm file when it is in use rather that only when opening the file with `Configuration.compactOnLaunch` enabled. ([Core Issue #5754](https://github.com/realm/realm-core/issues/5754))
 * Add support for `Realm.copyFromRealm()`. All RealmObjects, RealmResults, RealmList and RealmSets now also have a `copyFromRealm()` extension method.
 * [Sync] `App.close()` have been added so it is possible to close underlying ressources used by the app instance.
+* [Sync] Add support for progress listeners with `SyncSession.progress`. (Issue [#428](https://github.com/realm/realm-kotlin/issues/428))
 
 ### Fixed
 * Internal dispatcher threads would leak when closing Realms. (Issue [#818](https://github.com/realm/realm-kotlin/issues/818))
@@ -32,7 +33,7 @@ This release will bump the Realm file format from version 22 to 23. Opening a fi
 
 
 ### Internal
-* Updated to Realm Core 13.0.0, commit f68cf2eb795aac4a42700a2446d6b69fe89dd13b.
+* Updated to Realm Core 13.0.0, commit f3d022476fb28eef41763a333730331e67decd00.
 * Updated to require Swig 4.1.0.